successfully connects the backend to mongodb. Also changes backend to use es modules instead of commonjs - moves infisical code into its own module - uses conn.mjs to handle database connection logic - has the root api pathway return a tset query
