--- server/atlas_uri.js
@@ -1,1 +1,0 @@
-module.exports = uri = "<connection-string>";
